ci: gate the doc build and the macOS cfg arms
`cargo doc --no-deps --document-private-items` failed. The `Rules` docs link
to `Rules::none`, which a doc build never sees behind `#[cfg(test)]`. Widen
that gate to `any(test, doc)`. Drop two redundant explicit link targets, which
`-D warnings` rejects once private items are documented.

macOS carried six warnings no gate could catch. The QA job compiles only the
Linux arm of every `cfg(target_os)`, and the macOS job runs `make elle`, which
does not fail on warnings. Gate the bindings the io_uring backend alone
consumes. Narrow two dead-code allows to the platforms that compile
`io::uring` out. Fix three clippy lints that fire only there.

Add `make crosscheck`: clippy against x86_64-apple-darwin. Clippy neither
codegens nor links, so this needs no macOS SDK. It runs in the QA job and
before `make test`. The macOS job gains its own clippy step for `ffi` and
`jit`, which cannot cross-compile. Three jit-off warnings had to go first, or
the cross-check could not run under `-D warnings`.

`make test` now passes `--document-private-items` like CI does. Without it the
local gate cannot resolve a link into a `pub(crate)` item, so a broken one
reaches CI unseen.

+++ src/io/watch.rs
@@ -428,9 +428,9 @@ mod platform {
 
                 let path = inner.fd_to_path.get(&fd).cloned().unwrap_or_default();
 
-                let kind = if fflags & libc::NOTE_DELETE as u32 != 0 {
+                let kind = if fflags & libc::NOTE_DELETE != 0 {
                     WatchEventKind::Remove
-                } else if fflags & libc::NOTE_RENAME as u32 != 0 {
+                } else if fflags & libc::NOTE_RENAME != 0 {
                     WatchEventKind::Rename
                 } else {
                     // NOTE_WRITE, NOTE_EXTEND, NOTE_ATTRIB → Modify
